refactoring Contacts component and change background
