+++ src/components/Contact/Slack.js
@@ -1,0 +1,41 @@
+import React from 'react';
+import Globals from '../../utils/Globals';
+import { hover } from 'glamor';
+
+const Slack = ({ styles }) =>
+  <span>
+    <a
+      title="slack"
+      href={Globals.contacts.slack}
+      target="_blank"
+      rel="noopener noreferrer"
+    >
+      <svg
+        id="slack"
+        x="0px"
+        y="0px"
+        width="48px"
+        height="48px"
+        viewBox="0 0 31.444 31.443"
+      >
+        <g>
+          <path
+            fill="gray"
+            {...styles.transition}
+            {...hover({ fill: Globals.colors.primary })}
+            d="M31.202,16.369c-0.62-1.388-2.249-2.011-3.637-1.391l-1.325,0.594l-3.396-7.591l1.325-0.592
+                    c1.388-0.622,2.01-2.25,1.389-3.637c-0.62-1.389-2.248-2.012-3.637-1.39l-1.324,0.593l-0.593-1.326
+                    c-0.621-1.388-2.249-2.009-3.637-1.388c-1.388,0.62-2.009,2.247-1.389,3.637l0.593,1.325L7.98,8.598L7.388,7.273
+                    c-0.621-1.39-2.249-2.009-3.637-1.39C2.363,6.504,1.742,8.132,2.362,9.52l0.592,1.324L1.63,11.438
+                    c-1.388,0.621-2.01,2.247-1.389,3.636c0.62,1.388,2.249,2.01,3.637,1.39l1.325-0.594l3.394,7.592l-1.325,0.592
+                    c-1.388,0.621-2.009,2.25-1.389,3.637c0.621,1.389,2.249,2.011,3.637,1.391l1.324-0.593l0.593,1.325
+                    c0.621,1.389,2.249,2.01,3.637,1.389c1.387-0.62,2.009-2.248,1.388-3.636l-0.591-1.326l7.591-3.394l0.592,1.321
+                    c0.621,1.391,2.248,2.013,3.637,1.392c1.388-0.619,2.01-2.248,1.389-3.637l-0.592-1.324l1.323-0.594
+                    C31.201,19.384,31.823,17.757,31.202,16.369z M13.623,21.215l-3.395-7.593l7.591-3.394l3.395,7.591L13.623,21.215z"
+          />
+        </g>
+      </svg>
+    </a>
+  </span>;
+
+export default Slack;
